requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
altair==5.5.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
ccxt==4.4.80
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
fonttools==4.58.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
google-auth-oauthlib==1.2.2
gspread==6.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.25.2
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
multidict==6.4.3
narwhals==1.39.1
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.2.1
plotly==6.1.0
propcache==0.3.1
protobuf==6.31.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.8.0
pycoingecko==3.2.0
pycparser==2.22
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==20.7
pytz==2025.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
